pyproject.toml
[tool.poetry]
name = "interactive-llm-voice"
version = "0.1.0"
description = "diferent scripts to interact with LLMs using voice, new articles, and other data sources"
authors = ["Alonso Isidoro Román <alonsoir@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.28.0"
speechrecognition = "^3.10.4"
pyttsx3 = "^2.90"
pyaudio = "^0.2.14"
python-dotenv = "^1.0.1"
transformers = "^4.40.2"
black = "^24.4.2"
pyobjc = "^10.2"
pyobjc-framework-avfoundation = "^10.2"
scrapegraphai = "^0.10.0"
playwright = "^1.43.0"
dbfread = "^2.0.7"
pyproj = "^3.6.1"
pyarrow = "^16.0.0"
fastparquet = "^2024.2.0"
altair = "^5.3.0"
vega-datasets = "^0.9.0"
pandas = "^2.2.2"
ipython = "^8.24.0"
matplotlib = "^3.8.4"
seaborn = "^0.13.2"
plotly = "^5.22.0"
ipywidgets = "^8.1.2"
flask = "^3.0.3"
opencv-python = "^4.9.0.80"
moviepy = "^1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
